Time: 9 ms (29.26%), Space: 9.9 MB (12.65%) - LeetHub

+++ 0223-rectangle-area/0223-rectangle-area.cpp
@@ -12,6 +12,6 @@ public:
             toSub = (long long)x_dist * y_dist;
         }
 
-        return (int)(area1 + area2 - toSub);   
+        return (area1 + area2 - toSub);   
     }
 };
